refactor: UI->Component
