+++ app/src/main/java/com/example/jetprofile/components/DetailSection.kt
@@ -1,0 +1,36 @@
+package com.example.jetprofile.components
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Favorite
+import androidx.compose.material.icons.filled.LocationOn
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun DetailSection() {
+    Column (
+        modifier = Modifier
+            .fillMaxWidth()
+            .clip(RoundedCornerShape(10.dp))
+            .background(Color.LightGray.copy(alpha = 0.3f))
+            .padding(horizontal = 10.dp, vertical = 20.dp)
+    )
+    {
+        Label(
+            icon = Icons.Default.Favorite,
+            text = "趣味: 旅行, Zennの記事を書く",
+        )
+        Spacer(modifier = Modifier.height(10.dp))
+        Label(
+            icon = Icons.Default.LocationOn,
+            text = "居住地: 東京都",
+            color = Color.Gray,
+        )
+    }
+}
