+++ app/src/main/java/com/example/jetprofile/components/label.kt
@@ -1,0 +1,29 @@
+package com.example.jetprofile.components
+
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.width
+import androidx.compose.material.Icon
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+
+// アノテーションをつけるとJetpack Composeで使える
+@Composable
+fun Label(icon: ImageVector, text: String, color: Color = MaterialTheme.colors.onBackground) {
+    Row(verticalAlignment = Alignment.CenterVertically) {
+        Icon(
+            imageVector = icon,
+            contentDescription = null,
+        )
+        Spacer(modifier = Modifier.width(10.dp))
+        Text(text = text, color = color, fontSize = 14.sp, fontWeight = FontWeight.Bold)
+    }
+}
